+++ packages/ndk/lib/presentation_layer/init.dart
@@ -29,7 +29,7 @@ import '../domain_layer/usecases/jit_engine/jit_engine.dart';
 import '../domain_layer/usecases/lists/lists.dart';
 import '../domain_layer/usecases/lnurl/lnurl.dart';
 import '../domain_layer/usecases/metadatas/metadatas.dart';
-import '../domain_layer/usecases/nip05/verify_nip_05.dart';
+import '../domain_layer/usecases/nip05/nip_05.dart';
 import '../domain_layer/usecases/nwc/nwc.dart';
 import '../domain_layer/usecases/relay_manager.dart';
 import '../domain_layer/usecases/relay_sets/relay_sets.dart';
@@ -84,7 +84,7 @@ class Initialization {
   late FetchedRanges fetchedRanges;
   late ProofOfWork proofOfWork;
 
-  late VerifyNip05 verifyNip05;
+  late Nip05Usecase nip05;
 
   late final NetworkEngine engine;
 
@@ -107,6 +107,8 @@ class Initialization {
           accounts: accounts,
           nostrTransportFactory: _webSocketNostrTransportFactory,
           bootstrapRelays: _ndkConfig.bootstrapRelays,
+          eagerAuth: _ndkConfig.eagerAuth,
+          authCallbackTimeout: _ndkConfig.authCallbackTimeout,
         );
 
         engine = RelaySetsEngine(
@@ -123,6 +125,8 @@ class Initialization {
           nostrTransportFactory: _webSocketNostrTransportFactory,
           bootstrapRelays: _ndkConfig.bootstrapRelays,
           engineAdditionalDataFactory: JitEngineRelayConnectivityDataFactory(),
+          eagerAuth: _ndkConfig.eagerAuth,
+          authCallbackTimeout: _ndkConfig.authCallbackTimeout,
         );
 
         engine = JitEngine(
@@ -165,6 +169,7 @@ class Initialization {
       accounts: accounts,
       considerDonePercent: _ndkConfig.defaultBroadcastConsiderDonePercent,
       timeout: _ndkConfig.defaultBroadcastTimeout,
+      saveToCache: _ndkConfig.defaultBroadcastSaveToCache,
     );
 
     bunkers = Bunkers(
@@ -207,7 +212,7 @@ class Initialization {
       blockedRelays: _globalState.blockedRelays,
     );
 
-    verifyNip05 = VerifyNip05(
+    nip05 = Nip05Usecase(
       database: _ndkConfig.cache,
       nip05Repository: nip05repository,
     );