+++ packages/nip07_event_signer/CHANGELOG.md
@@ -1,3 +1,47 @@
+## 1.0.6-dev.11
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.10
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.9
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.8
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.7
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.6
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.5
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.4
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.3
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.2
+
+ - Update a dependency to the latest release.
+
+## 1.0.6-dev.1
+
+ - Update a dependency to the latest release.
+
 ## 1.0.6-dev.0
 
  - Update a dependency to the latest release.