+++ packages/rust_verifier/CHANGELOG.md
@@ -1,3 +1,47 @@
+## 0.5.0-dev.11
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.10
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.9
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.8
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.7
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.6
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.5
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.4
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.3
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.2
+
+ - Update a dependency to the latest release.
+
+## 0.5.0-dev.1
+
+ - Update a dependency to the latest release.
+
 ## 0.5.0-dev.0+1
 
  - Update a dependency to the latest release.